pyproject.toml
[project]
name = "wechat-login-mock"
version = "0.1.2"
description = ""
authors = []
requires-python = ">=3.9"
dependencies = [
    "djangorestframework (>=3.16.0)",
    "gunicorn (>=23.0.0)",
]

[tool.poetry.group.dev.dependencies]
fast-dev-cli = "0.15.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
